Altered SmarTaskController to add in new text area and remove hot fix
Altered the onEnter method of SmarTaskController to remove the temporary
hot fix for V0.1 testing and added in a more stable method for
taskWindow to display tasks for the day. Added a injection for the tasks
without a due date and a assert to check if the text area exists.